% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppDatiPuntuali.R
\name{AppDatiPuntuali}
\alias{AppDatiPuntuali}
\title{Applicazione per la visualizzazione di dati spazio temporali caso dati puntuali}
\usage{
AppDatiPuntuali()
}
\description{
Shiny app per la visualizzazione di dati spazio temporali.
Per maggiori informazioni riferirsi a:

https://github.com/AndreaMineo/VisualizzazioneDatiSpazioTemporali
}
