% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppDatiRegionali.R
\name{AppDatiRegionali}
\alias{AppDatiRegionali}
\title{Applicazione per la visualizzazione di dati spazio temporali caso dati regionali}
\usage{
AppDatiRegionali()
}
\description{
Shiny app per la visualizzazione di dati spazio temporali.
Per maggiori informazioni riferirsi a:

https://github.com/AndreaMineo/VisualizzazioneDatiSpazioTemporali
}
