% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizzazioneDatiAreali.R
\name{VisualizzazioneDatiAreali}
\alias{VisualizzazioneDatiAreali}
\title{Applicazione per la visualizzazione di dati spazio temporali caso dati regionali}
\usage{
VisualizzazioneDatiAreali()
}
\description{
Shiny app per la visualizzazione di dati spazio temporali.
Per maggiori informazioni riferirsi a:

https://github.com/AndreaMineo/VisualizzazioneDatiSpazioTemporali
}
